requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
setuptools==69.5.1
sqlparse==0.5.0
uritemplate==4.1.1
wheel==0.43.0
